refactor: examples move to esm
