+++ examples/event/Main.ts
@@ -1,7 +1,7 @@
 import "reflect-metadata";
-import { dirname, importx } from "../../packages/importer/src/esm/index.js";
-import { Client } from "../../src/index.js";
+import { Client } from "../../build/cjs/index.cjs";
 import { Intents } from "discord.js";
+import { importx } from "../../packages/importer/build/cjs/index.cjs";
 
 export class Main {
   private static _client: Client;
@@ -25,7 +25,7 @@ export class Main {
       console.log("Bot started...");
     });
 
-    await importx(dirname(import.meta.url) + "/discords/**/*.{js,ts}");
+    await importx(__dirname + "/discords/**/*.{js,ts}");
     await this._client.login(process.env.BOT_TOKEN ?? "");
   }
 }
