+++ examples/event/discords/AppDiscord.ts
@@ -1,5 +1,5 @@
-import { Client, Discord, On } from "../../../src/index.js";
-import type { ArgsOf } from "../../../src/index.js";
+import { Client, Discord, On } from "../../../build/cjs/index.cjs";
+import type { ArgsOf } from "../../../build/cjs/index.cjs";
 
 @Discord()
 export abstract class AppDiscord {
