+++ examples/guards/guards/NotBot.ts
@@ -2,7 +2,7 @@ import {
   ArgsOf,
   GuardFunction,
   SimpleCommandMessage,
-} from "../../../src/index.js";
+} from "../../../build/cjs/index.cjs";
 import {
   ButtonInteraction,
   CommandInteraction,
