+++ examples/slash/discords/Permission.ts
@@ -5,7 +5,7 @@ import {
   Permission,
   Slash,
   SlashOption,
-} from "../../../src/index.js";
+} from "../../../build/cjs/index.cjs";
 
 @Discord()
 export abstract class AppDiscord {
