--- examples/slash/package.json
@@ -1,6 +1,0 @@
-{
-  "type": "module",
-  "scripts": {
-    "start": "node --loader ts-node/esm/transpile-only Main.ts"
-  }
-}
