+++ examples/tsyringe/Main.ts
@@ -1,8 +1,8 @@
 import "reflect-metadata";
-import { Client, DIService } from "../../src/index.js";
-import { dirname, importx } from "../../packages/importer/src/esm/index.js";
+import { Client, DIService } from "../../build/cjs/index.cjs";
 import { Intents } from "discord.js";
 import { container } from "tsyringe";
+import { importx } from "../../packages/importer/build/cjs/index.cjs";
 
 // initialize tsyringe container
 // it's important, this done before calling bot.login
@@ -32,7 +32,7 @@ export class Main {
       this._client.executeInteraction(interaction);
     });
 
-    await importx(dirname(import.meta.url) + "/discords/**/*.{js,ts}");
+    await importx(__dirname + "/discords/**/*.{js,ts}");
     await this._client.login(process.env.BOT_TOKEN ?? "");
   }
 }
