+++ examples/tsyringe/discords/AppDiscord.ts
@@ -1,4 +1,4 @@
-import { DIService, Discord, Slash } from "../../../src/index.js";
+import { DIService, Discord, Slash } from "../../../build/cjs/index.cjs";
 import { container, injectable, singleton } from "tsyringe";
 import { CommandInteraction } from "discord.js";
